Rusty Skylines Alpha v1.3.5a Dec 14th 23:44
z_near: 5.0 now instead of 0.1

Game can be closed by simply pressing Escape, I added the saving settings.toml feature so it saves when closing the game as well!

Changed ALL noise in the terrain generator from uselessly precise "scientific" f64 precision to just f32, which is plenty for this game. That means WAY better performance!
For nerds: I switched from Fbm::<Perlin> to FastNoiseLite!

Removed time stuff from the sky uniforms, because of this:
Contrary to downgrading noise accuracy, I increased the accuracy of the fabric of time itself, to f64, so now this game will never get glitchy time stuff overflow whatever! (Or you wait thousands or millions of years, I doubt you have money to pay the electricity bill for yur pc for that long!)

I am saving and loading the in game time from settings.toml now.

Improved the color of terrain a teeny tiny bit. Or not? Idk maybe I'm Joe Biden and have dementia.  Ah no that was yesterday but I haven't committed it yet!

+++ src/renderer/ui.rs
@@ -224,7 +224,7 @@ impl UiRenderer {
         }
         let new_uniform = ScreenUniform {
             size: [size.0, size.1],
-            time: time.total_time,
+            time: time.total_time as f32,
             enable_dither: 1,
             mouse: mouse.pos.to_array(),
         };
@@ -1003,7 +1003,7 @@ impl UiRenderer {
                     let y1 = tp.pos[1] + atlas.line_height + caret_offset_y;
 
                     let t = time_system.total_time * 3.0; // adjust speed here
-                    let blink_alpha = 0.5 + 0.5 * t.cos(); // smooth
+                    let blink_alpha = (0.5 + 0.5 * t.cos()) as f32; // smooth
                     let caret_alpha = blink_alpha.clamp(0.0, 1.0);
 
                     let caret_color = [1.0, 1.0, 1.0, caret_alpha];
@@ -1446,7 +1446,7 @@ impl UiRenderer {
                     let y1 = tp.pos[1] + atlas.line_height + caret_offset_y;
 
                     let t = time_system.total_time * 3.0; // adjust speed here
-                    let blink_alpha = 0.5 + 0.5 * t.cos(); // smooth
+                    let blink_alpha = (0.5 + 0.5 * t.cos()) as f32; // smooth
                     let caret_alpha = blink_alpha.clamp(0.0, 1.0);
 
                     let caret_color = [1.0, 1.0, 1.0, caret_alpha];
